+++ docs/misc/changelog.rst
@@ -24,6 +24,7 @@ New Features:
 - Added ``env_is_wrapped()`` method for ``VecEnv`` to check if its environments are wrapped
   with given Gym wrappers.
 - Added ``monitor_kwargs`` parameter to ``make_vec_env`` and ``make_atari_env``
+- Wrap the environments automatically with a ``Monitor`` wrapper when possible.
 
 Bug Fixes:
 ^^^^^^^^^^